boards: Update nrf5340dk_nrf5340 to HWMv2 scheme
This commit updates nrf5340dk_nrf5340 to use board scheme v2, and
thereby remove the Kconfig BOARD setting, as this is now passed from
CMake.

The nRF5340dk now support cpuapp and cpunet as cpusets, and the ns as
a build variant.

It also uses the new SoC approach, and thereby ensuring that the right
SoC is always used, and cannot accidentially be changed by the user.

Signed-off-by: Jamie McCrae <jamie.mccrae@nordicsemi.no>